==> 3_vectorize_make_dict_steps.ipynb <==
## ✫paraphrase-multilingual-MiniLM-L12-v2を使用して文章のベクトル化

### 1. 文章の本文をベクトル化するためにparaphrase-multilingual-MiniLM-L12-v2をロード

import sentence_transformers
from sentence_transformers import SentenceTransformer

# https://huggingface.co/sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2
# This is a sentence-transformers model: It maps sentences & paragraphs to a 384 dimensional dense vector space and can be used for tasks like clustering or semantic search.
sbert = SentenceTransformer('paraphrase-multilingual-MiniLM-L12-v2')
vector = sbert.encode('100回分のループを渡しているので, 1回のループで1%の進捗があります.')

vector.shape
# (384,)

### 2. 文章の本文をリスト化

import os
# カテゴリーのフォルダのみを抽出
categories = [name for name in os.listdir(
    "./data/livedoor/text/") if os.path.isdir("./data/livedoor/text/"+name)]

def extract_main_txt(file_name):
    with open(file_name) as text_file:
        # 今回はタイトル行は外したいので、3要素目以降の本文のみ使用
        text = text_file.readlines()[3:]

        # 3要素目以降にも本文が入っている場合があるので、リストにして、後で結合させる
        text = [sentence.strip() for sentence in text]  # 空白文字(スペースやタブ、改行)の削除
        text = list(filter(lambda line: line != '', text))
        text = ''.join(text)
        text = text.translate(str.maketrans(
            {'\n': '', '\t': '', '\r': '', '\u3000': ''}))  # 改行やタブ、全角スペースを消す
        return text

# リストに前処理した本文と、カテゴリーのラベルを追加していく
import glob

list_text = []
list_label = []

for cat in categories:
    text_files = glob.glob(os.path.join("./data/livedoor/text", cat, "*.txt"))

    # 前処理extract_main_txtを実施して本文を取得
    body = [extract_main_txt(text_file) for text_file in text_files]

    label = [cat] * len(body)  # bodyの数文だけカテゴリー名のラベルのリストを作成

    list_text.extend(body)  # appendが要素を追加するのに対して、extendはリストごと追加する
    list_label.extend(label)

### 3. Livedoorテキストの各本文をベクトル化して、以下の3つを作成する
- 文章ベクトル → 文章本文の辞書
- （文章ID, 文章ベクトル）を配列としたリスト
- 文章本文 → 文章ラベルの辞書

import numpy as np
livedoor_vec_to_text_dict = {}
livedoor_vec_list = []
livedoor_text_to_label_dict = {}

from tqdm import tqdm 

for text_item, label_item in tqdm(zip(list_text, list_label)):
    vector = sbert.encode(text_item)
    n = np.array(vector)
    n = tuple(n)
    livedoor_vec_to_text_dict[n] = text_item
    livedoor_vec_list.append((len(livedoor_vec_list), n))
    livedoor_text_to_label_dict[text_item] = label_item

print(len(list_text))
print(len(livedoor_vec_to_text_dict))
print(len(livedoor_vec_list))
print(len(livedoor_text_to_label_dict))
# 7376
# 7365 → 同じものがあるのか？
# 7376
# 7368 → 同じものがあるのか？

# 作成した辞書（文章ベクトル→文章本文）の検証
vector = sbert.encode(list_text[10])
n = np.array(vector)
n = tuple(n)
livedoor_vec_to_text_dict[n]

# 10番目の文章とラベルを確認
print(list_text[10])
print(list_label[10])

# 作成した辞書（文章ベクトル→文章本文）の検証
vector = sbert.encode(list_text[1015])
n = np.array(vector)
n = tuple(n)
livedoor_vec_to_text_dict[n]

# 15番目の文章とラベルを確認
print(list_text[1015])
print(list_label[1015])

### 4. 作成した辞書の永続化

import pickle
with open("livedoor_vec_to_text_dict.pkl","wb") as f:
    pickle.dump(livedoor_vec_to_text_dict, f)

with open("livedoor_text_to_label_dict.pkl","wb") as f:
    pickle.dump(livedoor_text_to_label_dict, f)   

### 5. ベクトルのリストの永続化

import pickle
with open("livedoor_vec_list.pkl","wb") as f:
    pickle.dump(livedoor_vec_list, f)
